+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##  [0.9.0] - 2025-10-27
+
+- Make changes to controller idempotent for both go and python
+- adding tests and lint to precommit hook.
+
 ##  [0.8.0] - 2025-10-27
 
 - When change python controller, make app shut down previous action safely.